set up gmx fep
